requirements.txt
python-dotenv==1.1.1
python_fasthtml==0.12.22
supabase==2.17.0
